Further separate main branch patches from older branches

+++ run.sh
@@ -50,33 +50,40 @@ if [[ ! -d "${PYTHON_DIR}/inst/${PYTHON_VER}" ]]; then
     cd "${PROJECT_DIR}"
 fi
 
+cd ${PYTHON_DIR}
+
 # Configure Python build. Python 3.11+ autodetects modules.
 export CONFIG_SITE="${PROJECT_DIR}/config.site"
 if [[ ("$PYTHON_MAJOR" -ge "3") && ("$PYTHON_MINOR" -ge "11") ]]; then
     rm -f "${PYTHON_DIR}/Modules/Setup.local"
+    patch -p1 -N -r- < ${PROJECT_DIR}/patches/getpath.py.patch
 else
-    cp "Setup.local" "${PYTHON_DIR}/Modules/Setup.local"
-fi
+    cp "${PROJECT_DIR}/Setup.local" "${PYTHON_DIR}/Modules/Setup.local"
 
-cd ${PYTHON_DIR}
+    # Apply patches
+    patch -p1 -N -r- < ${PROJECT_DIR}/patches/configure.ac.patch
 
-# Apply patches
-patch -p1 -N -r- < ${PROJECT_DIR}/patches/configure.ac.patch
+    if [[ -f "${PYTHON_DIR}/Modules/_zoneinfo.c" ]]; then
+        patch -p1 -N -r- < ${PROJECT_DIR}/patches/_zoneinfo.c.patch
+    fi
 
-if [[ -f "${PYTHON_DIR}/Modules/_zoneinfo.c" ]]; then
-    patch -p1 -N -r- < ${PROJECT_DIR}/patches/_zoneinfo.c.patch
-fi
-
-if [[ ("$PYTHON_MAJOR" -eq "3") && ("$PYTHON_MINOR" -le "8") ]]; then
-    sed -i 's/_zoneinfo/#_zoneinfo/' "${PYTHON_DIR}/Modules/Setup.local"
-    sed -i 's/_decimal/#_decimal/' "${PYTHON_DIR}/Modules/Setup.local"
+    if [[ ("$PYTHON_MAJOR" -eq "3") && ("$PYTHON_MINOR" -le "8") ]]; then
+        sed -i 's/_zoneinfo/#_zoneinfo/' "${PYTHON_DIR}/Modules/Setup.local"
+        sed -i 's/_decimal/#_decimal/' "${PYTHON_DIR}/Modules/Setup.local"
+    fi
 fi
 
 # Set compiler flags
 export CC="clang --target=wasm32-wasi"
-export CFLAGS="-g -D_WASI_EMULATED_GETPID -D_WASI_EMULATED_SIGNAL -I/opt/include -I${WASIX_DIR}/include -isystem ${WASIX_DIR}/include"
+export CFLAGS="-g -D_WASI_EMULATED_GETPID -D_WASI_EMULATED_SIGNAL -D_WASI_EMULATED_PROCESS_CLOCKS -I/opt/include -I${WASIX_DIR}/include -isystem ${WASIX_DIR}/include"
+export CPPFLAGS="${CFLAGS}"
 export LIBS="-Wl,--stack-first -Wl,-z,stack-size=83886080 -L/opt/lib -L${WASIX_DIR} -lwasix -lwasi-emulated-signal"
-export PATH=${PYTHON_DIR}/inst/${PYTHON_VER}/bin:${PATH}
+export PATH=${PYTHON_DIR}/inst/${PYTHON_VER}/bin:${PROJECT_DIR}/build/bin:${PATH}
+
+# Override ld. This is called to build _ctype_test as a "shared module" which isn't supported.
+mkdir -p "${PROJECT_DIR}/build/bin"
+echo "wasm-ld --stack-first -z,stack-size=83886080 -L/opt/lib -L${WASIX_DIR} -lwasix -L${WASI_SDK_PATH}/share/wasi-sysroot/lib/wasm32-wasi/ -lwasi-emulated-signal --no-entry \$*" > "${PROJECT_DIR}/build/bin/ld"
+chmod +x "${PROJECT_DIR}/build/bin/ld"
 
 # Configure and build
 cp ${WASI_SDK_PATH}/share/misc/config.sub . && \
@@ -97,16 +104,18 @@ rm -f "${PYTHON_DIR}/Modules/Setup.local"
 cd ${PROJECT_DIR}
 
 # Package wasi-python and wasix libraries
-tar zcvf ${PROJECT_DIR}/wasi-python.tgz /opt/wasi-python
-mkdir -p /opt/wasix/lib && \
-    cp -R ${WASIX_DIR}/include /opt/wasix/. && \
-    cp ${WASIX_DIR}/libwasix.a /opt/wasix/lib/. && \
-    tar zcvf ${PROJECT_DIR}/wasix.tgz /opt/wasix
+if [[ -f "/opt/wasi-python/bin/python${PYTHON_VER}.wasm" ]]; then
+    tar zcvf ${PROJECT_DIR}/wasi-python.tgz /opt/wasi-python
 
-# This is needed when running unit tests.
-mkdir -p tmp
+    mkdir -p /opt/wasix/lib && \
+        cp -R ${WASIX_DIR}/include /opt/wasix/. && \
+        cp ${WASIX_DIR}/libwasix.a /opt/wasix/lib/. && \
+        tar zcvf ${PROJECT_DIR}/wasix.tgz /opt/wasix
 
-# Reality check it
-wasmtime run --mapdir=/opt/wasi-python::/opt/wasi-python \
-	     --env PATH=/opt/wasi-python/bin \
-	     -- /opt/wasi-python/bin/python${PYTHON_VER}.wasm -V
+    # Reality check it
+    wasmtime run --mapdir=/opt/wasi-python::/opt/wasi-python \
+    	         --env PATH=/opt/wasi-python/bin \
+    	         -- /opt/wasi-python/bin/python${PYTHON_VER}.wasm -V
+else
+    echo "ERROR: No Python build was found."
+fi
